OrderPanel searchBar and confirmoffer
Orderpanel searchtextpanel added and added the functions to it. *Tested*
Confirming the offer by clicking on the button is started but failed. Now button has no function
MainMenu populateclasses extended with order creation
Populateclasses accesslevel set to public
MainMenu imported to the OrderPanel, so the populateclasses method is usable there

Co-Authored-By: WrexBG <40688355+WrexBG@users.noreply.github.com>